v0.1.0 Proof of concept to first release like a Library.
