+++ README.md
@@ -1,8 +1,8 @@
-# esp32s3 camera tests
+# ESPNowCam - ESP32 Camera vis ESPNow
 
 [![PlatformIO](https://github.com/hpsaturn/esp32s3-cam/workflows/PlatformIO/badge.svg)](https://github.com/hpsaturn/esp32s3-cam/actions/) ![ViewCount](https://views.whatilearened.today/views/github/hpsaturn/esp32s3-cam.svg)  
 
-Basic tests and examples over ESP32S3 Cameras.
+ESPNowCam library, is a straightforward video streamer for popular ESP32Cam models, leveraging the ESPNow protocol. No need for IPs, routers, or credentials—keeping it simple! :D
 
 <table>
   <tr>
@@ -25,25 +25,19 @@ Basic tests and examples over ESP32S3 Cameras.
 
 - [x] M5Core2 (AWS tested)
 - [x] M5CoreS3
-- [ ] Maybe any TFT with LGFX support and PSRAM
+- [x] Makerfabs Parallel using LGFX
+- [ ] Maybe any TFT with LGFX support (better with PSRAM)
 
 ## Examples and Tests
 
 | ENV Name   |    Target      |  Status |
 |:-----------------|:--------------:|:----------:|
-| freenove-basic   |  Basic FPS and JPEG speed tester | STABLE |
-| freenove-espnow  | ESPNow camera transmitter | STABLE |
-| freenove-espnow-stats | Basic stats via ESPNowJoystick [1] | STABLE |
-| freenove-webcam | Original Freenove example with improvements  | STABLE |
-| m5cores3-basic | Basic builtin Camera/Display test | STABLE |
-| m5cores3-espnow | ESPNow built-in camera transmitter | STABLE |
-| m5cores3-espnow-stats |Basic stats via ESPNowJoystick [1] | STABLE |
+| freenove-espnow-sender  | ESPNow camera transmitter | STABLE |
+| m5cores3-espnow-sender | ESPNow built-in camera transmitter | STABLE |
 | m5cores3-espnow-receiver | Video receiver via ESPNow [2] | STABLE|
-| core2-espnow-receiver | Video receiver via ESPNow  | STABLE |
+| m5core2-espnow-receiver | Video receiver via ESPNow [2] | STABLE |
 
-[1] For receive the stats, configure other ESP32 how is explained here: [ESPNowJoystick Library](https://github.com/hpsaturn/espnow-joystick#readme)
-
-[2] Use with freenove or M5CoreS3 espnow sample
+[2] Use with freenove or M5CoreS3 espnow sender sample
 
 ## Usage and Install
 
@@ -59,9 +53,9 @@ pio run -e m5cores3-espnow-receiver --target upload
 
 - [x] NanoPb possible issue #1 (payload size)
 - [x] Unified ESPNow in an one class for all transmitters and receivers
+- [x] Isolate the ESPNow Receiver and Transmitter in a seperated library
 - [ ] Add sender callback to improve speed
-- [ ] Isolate the ESPNow Receiver and Transmitter in a seperated library
-- [ ] Migration to esp_wifi_80211_tx() to improve Payload
+- [ ] Migration to esp_wifi_80211_tx() to improve Payload and Quality
 
 ## ESPNow Transmitter and Receiver
 
@@ -71,8 +65,6 @@ The last version has many improvements, and right now is stable. It needs some t
 
 ## Troubleshooting
 
-The Freenove camera sometimes needs good power cable and also takes some seconds to stabilization, that means, that not worries for initial crashes.
-
-**This is a proof of concept and some basic tests and examples, don't hope to much**  
+The Freenove camera sometimes needs good power cable and also takes some seconds to stabilization, that means, that not worries for initial video glitches.
 
 ---
