--- examples/freenove-basic/freenove-basic.cpp
@@ -1,53 +1,0 @@
-#include <Arduino.h>
-#include <WiFi.h>
-#include <CamFreenove.h>
-
-// #define CONVERT_TO_JPEG
-
-CamFreenove Camera;
-
-void setup() {
-  Serial.begin(115200);
-  Serial.setDebugOutput(true);
-  Serial.println();
-  delay(1000);
-
-  if (!Camera.begin()) {
-    Serial.println("Camera Init Fail");
-  }
-  Camera.sensor->set_framesize(Camera.sensor, FRAMESIZE_QVGA); 
-
-  if(psramFound()){
-    size_t psram_size = esp_spiram_get_size() / 1048576;
-    Serial.printf("PSRAM size: %dMb\r\n", psram_size);
-  }
-}
-
-uint16_t frame = 0;
-
-void printFPS(const char *msg) {
-  static uint_least64_t timeStamp = 0;
-  frame++;
-  if (millis() - timeStamp > 1000) {
-    timeStamp = millis();
-    Serial.printf("%s %2d FPS\r\n",msg, frame);
-    frame = 0;
-  } 
-}
-
-void loop() {
-  if (Camera.get()) {
-#ifdef CONVERT_TO_JPEG
-    uint8_t *out_jpg = NULL;
-    size_t out_jpg_len = 0;
-    frame2jpg(Camera.fb, 64, &out_jpg, &out_jpg_len);
-    printFPS("JPG compression at");
-    // Display.drawJpg(out_jpg, out_jpg_len, 0, 0, dw, dh);
-    free(out_jpg);
-#else
-    printFPS("frame ready at");
-    // Display.pushImage(0, 0, dw, dh, (uint16_t *)CoreS3.Camera.fb->buf);
-#endif
-    Camera.free();
-  }
-}
