--- examples/freenove-webcam/freenove-webcam.cpp
@@ -1,121 +1,0 @@
-/**********************************************************************
-  Filename    : Camera Web Server
-  Description : The camera images captured by the ESP32S3 are displayed on the web page.
-  Auther      : www.freenove.com
-  Modification: 2022/10/31
-**********************************************************************/
-#include "esp_camera.h"
-#include <WiFi.h>
-
-// ===================
-// Select camera model
-// ===================
-//#define CAMERA_MODEL_WROVER_KIT // Has PSRAM
-//#define CAMERA_MODEL_ESP_EYE // Has PSRAM
-#define CAMERA_MODEL_ESP32S3_EYE // Has PSRAM
-//#define CAMERA_MODEL_M5STACK_PSRAM // Has PSRAM
-//#define CAMERA_MODEL_M5STACK_V2_PSRAM // M5Camera version B Has PSRAM
-//#define CAMERA_MODEL_M5STACK_WIDE // Has PSRAM
-//#define CAMERA_MODEL_M5STACK_ESP32CAM // No PSRAM
-//#define CAMERA_MODEL_M5STACK_UNITCAM // No PSRAM
-//#define CAMERA_MODEL_AI_THINKER // Has PSRAM
-//#define CAMERA_MODEL_TTGO_T_JOURNAL // No PSRAM
-// ** Espressif Internal Boards **
-//#define CAMERA_MODEL_ESP32_CAM_BOARD
-//#define CAMERA_MODEL_ESP32S2_CAM_BOARD
-//#define CAMERA_MODEL_ESP32S3_CAM_LCD
-
-#include "camera_pins.h"
-
-// ===========================
-// Enter your WiFi credentials
-// ===========================
-
-const char *ssid =     WIFI_SSID;         // Put your SSID here
-const char *password = WIFI_PASS;     // Put your PASSWORD here
-
-void startCameraServer();
-
-void setup() {
-  Serial.begin(115200);
-  Serial.setDebugOutput(true);
-  Serial.println();
-
-  camera_config_t config;
-  config.pin_pwdn = PWDN_GPIO_NUM;
-  config.pin_reset = RESET_GPIO_NUM;
-  config.pin_xclk = XCLK_GPIO_NUM;
-  config.pin_sccb_sda = SIOD_GPIO_NUM;
-  config.pin_sccb_scl = SIOC_GPIO_NUM;
-  config.pin_d7 = Y9_GPIO_NUM;
-  config.pin_d6 = Y8_GPIO_NUM;
-  config.pin_d5 = Y7_GPIO_NUM;
-  config.pin_d4 = Y6_GPIO_NUM;
-  config.pin_d3 = Y5_GPIO_NUM;
-  config.pin_d2 = Y4_GPIO_NUM;
-  config.pin_d1 = Y3_GPIO_NUM;
-  config.pin_d0 = Y2_GPIO_NUM;
-  
-  config.pin_vsync = VSYNC_GPIO_NUM;
-  config.pin_href = HREF_GPIO_NUM;
-  config.pin_pclk = PCLK_GPIO_NUM;
-
-  config.xclk_freq_hz = 20000000;
-  config.ledc_timer = LEDC_TIMER_0;
-  config.ledc_channel = LEDC_CHANNEL_0;
-
-  config.pixel_format = PIXFORMAT_JPEG; // for streaming
-  config.frame_size = FRAMESIZE_UXGA;
-  config.jpeg_quality = 12;
-  config.fb_count = 1;
-  config.fb_location = CAMERA_FB_IN_PSRAM;
-  config.grab_mode = CAMERA_GRAB_WHEN_EMPTY;
-  
-  // if PSRAM IC present, init with UXGA resolution and higher JPEG quality
-  // for larger pre-allocated frame buffer.
-  if(psramFound()){
-    size_t psram_size = esp_spiram_get_size() / 1048576;
-    Serial.printf("PSRAM size: %dMb\r\n", psram_size);
-    config.jpeg_quality = 10;
-    config.fb_count = 2;
-    config.grab_mode = CAMERA_GRAB_LATEST;
-  } else {
-    // Limit the frame size when PSRAM is not available
-    config.frame_size = FRAMESIZE_SVGA;
-    config.fb_location = CAMERA_FB_IN_DRAM;
-  }
-
-  // camera init
-  esp_err_t err = esp_camera_init(&config);
-  if (err != ESP_OK) {
-    Serial.printf("Camera init failed with error 0x%x", err);
-    return;
-  }
-
-  sensor_t * s = esp_camera_sensor_get();
-  // initial sensors are flipped vertically and colors are a bit saturated
-  s->set_vflip(s, 1); // flip it back
-  s->set_brightness(s, 1); // up the brightness just a bit
-  s->set_saturation(s, 0); // lower the saturation
-  
-  WiFi.begin(ssid, password);
-  WiFi.setSleep(false);
-
-  while (WiFi.status() != WL_CONNECTED) {
-    delay(500);
-    Serial.print(".");
-  }
-  Serial.println("");
-  Serial.println("WiFi connected");
-
-  startCameraServer();
-
-  Serial.print("Camera Ready! Use 'http://");
-  Serial.print(WiFi.localIP());
-  Serial.println("' to connect");
-}
-
-void loop() {
-  // Do nothing. Everything is done in another task by the web server
-  delay(10000);
-}
