+++ library.properties
@@ -1,0 +1,10 @@
+name=EspNowCam
+version=0.1.0
+author=@hpsaturn
+maintainer=Antonio Vanegas <hpsaturn@gmail.com>
+sentence=ESPNowCam, a straightforward video streamer for popular ESP32Cam models, leveraging the ESPNow protocol. No need for IPs, routers, or credentials—keeping it simple! :D
+paragraph=ESPNowCam supports ESP32 Cameras for give an easy video streamer from popular ESP32Cam models and it uses ESPNow protocol. No IP, No routers, No credentials. It has broadcast and P2P support. Tested on ESP32S3 cameras. Please visit my ESP32 Telegram channel https://t.me/esp_es and ask any doubt.
+category=Device Control
+url=https://github.com/hpsaturn/esp32s3-cam
+architectures=esp32
+license=GPL-3.0-only
