+++ platformio.ini
@@ -4,6 +4,7 @@
 
 [platformio]
 src_dir = ./examples/
+lib_dir = ./
 
 [wifi]
 ssid = ${sysenv.PIO_WIFI_SSID}         
@@ -29,31 +30,12 @@ board = esp32-s3-devkitc-1
 board_build.flash_size = 16MB
 board_build.partitions = ./config/partitions.csv
 
-[env:freenove-webcam]
+[env:freenove-espnow-sender]
 extends = esp32common
 board_build.arduino.memory_type = dio_opi ;
-build_src_filter = -<*> +<freenove-webcam/>
-
-[env:freenove-basic]
-extends = esp32common
-board_build.arduino.memory_type = dio_opi ;
-build_src_filter = -<*> +<freenove-basic/>
+build_src_filter = -<*> +<freenove-espnow-sender/>
 lib_deps = 
   nanopb/Nanopb@0.4.8
-
-[env:freenove-espnow]
-extends = esp32common
-board_build.arduino.memory_type = dio_opi ;
-build_src_filter = -<*> +<freenove-espnow/>
-lib_deps = 
-  nanopb/Nanopb@0.4.8
-
-[env:freenove-espnow-stats]
-extends = esp32common
-board_build.arduino.memory_type = dio_opi ;
-build_src_filter = -<*> +<freenove-espnow-stats/>
-lib_deps = 
-  hpsaturn/EspNowJoystick@^0.1.5
 
 [m5cores3_common]
 extends = esp32common
@@ -66,23 +48,12 @@ build_flags =
   -DARDUINO_USB_CDC_ON_BOOT=1
   -DARDUINO_USB_MODE=1
 
-[env:m5cores3-basic]
+[env:m5cores3-espnow-sender]
 extends = m5cores3_common
-build_src_filter = -<*> +<m5cores3-basic/>
-
-[env:m5cores3-espnow]
-extends = m5cores3_common
-build_src_filter = -<*> +<m5cores3-espnow/>
+build_src_filter = -<*> +<m5cores3-espnow-sender/>
 lib_deps =
   ${m5cores3_common.lib_deps}
   nanopb/Nanopb@0.4.8
-
-[env:m5cores3-espnow-stats]
-extends = m5cores3_common
-build_src_filter = -<*> +<m5cores3-espnow-stats/>
-lib_deps =
-  ${m5cores3_common.lib_deps}
-  hpsaturn/EspNowJoystick@^0.1.5
 
 [env:m5cores3-espnow-receiver]
 extends = m5cores3_common
@@ -91,17 +62,11 @@ lib_deps =
   ${m5cores3_common.lib_deps}
   nanopb/Nanopb@0.4.8
 
-[env:core2-espnow-receiver]
+[env:m5core2-espnow-receiver]
 extends = esp32common
 board = esp32dev
-build_src_filter = -<*> +<core2-espnow-receiver/>
+build_src_filter = -<*> +<m5core2-espnow-receiver/>
 lib_deps =
   m5stack/M5Unified@^0.1.6
   nanopb/Nanopb@0.4.8
 
-[env:makerfabs-espnow-receiver]
-extends = esp32common
-build_src_filter = -<*> +<makerfabs-espnow-receiver/>
-lib_deps =
-	lovyan03/LovyanGFX@^1.1.5
-  nanopb/Nanopb@0.4.8
