+++ src/CamFreenove.cpp
